WeatherSettings moved to package com.datastax.weather
