+++ weather-project/weather-app/src/main/resources/application.conf
@@ -6,7 +6,6 @@
 killrweather {
 
   app-name = "KillrWeather"
-##  spark.checkpoint.dir = "./tmp"
 
   cassandra {
     keyspace = "isd_weather_data"
