+++ weather-project/weather-app/src/test/scala/com/datastax/killrweather/WeatherSettingsSpec.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.datastax.killrweather
+package com.datastax.weather
 
 class SettingsSpec extends AbstractSpec with TestFileHelper {
